+++ CHANGELOG.md
@@ -1,17 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## v4.1.1 - 2026-03-25
+## v4.2 - 2026-04-28
 
 ### Improvements
 - Update the email alert template to cut the beginning of the last ping body (#939)
 - Update the email integration to attach email pings as .eml attachments
 
 ### Bug Fixes
-- Fix sendalerts to avoid crashing on a timed out db connection (#1267)
-- Fix the update_timeout view to also kick off nags
 - Fix the email integration to handle deep attachment stacks in last ping body
 - Fix check name and tag escaping in the Shell integration
+
+## v4.1.1 - 2026-03-25
+
+### Bug Fixes
+- Fix sendalerts to avoid crashing on a timed out db connection (#1267)
+- Fix the update_timeout view to also kick off nags
 
 ## v4.1 - 2026-03-20